Disable engine-strict option
We need to temporarily disable the engine-strict setting to unblock the UI
release pipeline: there are currently changes pending release older than the
set of changes where the node versioning was made strict here and in the UI
repo. Those changes need to go through the entire pipeline again in order to be
releasable, as certain temporary pipeline files required for the release have
expired and they need to be regenerated. Once the releases are through, we will
re-enable engine-strict here.